created file AppUserSignInService to implement user signin authentication, did this to resolve some cyclic dependencies
