+++ src/main/java/com/clientService/user/service/AppUserService.java
@@ -10,9 +10,11 @@ import com.clientService.user.repository.AppUserRepository;
 import com.clientService.user.repository.PortfolioRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
+@Service
 public class AppUserService {
     private final AppUserRepository appUserRepository;
     private final HttpResponseBody response;
